Remove and refactor
