--- src/EditorFeatures/CSharpTest/Diagnostics/Async/AddAwaitTests.cs
@@ -1,20 +1,0 @@
-﻿//// Licensed to the .NET Foundation under one or more agreements.
-//// The .NET Foundation licenses this file to you under the MIT license.
-//// See the LICENSE file in the project root for more information.
-
-//#nullable disable
-
-//using System.Threading.Tasks;
-//using Microsoft.CodeAnalysis.Test.Utilities;
-//using Xunit;
-//using VerifyCS = Microsoft.CodeAnalysis.Editor.UnitTests.CodeActions.CSharpCodeFixVerifier<
-//    Microsoft.CodeAnalysis.Testing.EmptyDiagnosticAnalyzer,
-//    Microsoft.CodeAnalysis.CSharp.CodeFixes.Async.CSharpAddAwaitCodeFixProvider>;
-
-//namespace Microsoft.CodeAnalysis.Editor.CSharp.UnitTests.Diagnostics.Async
-//{
-//    public class AddAwaitTests
-//    {
-
-//    }
-//}
