--- src/EditorFeatures/VisualBasicTest/Diagnostics/Async/AddAwaitTests.vb
@@ -1,17 +1,0 @@
-﻿'' Licensed to the .NET Foundation under one or more agreements.
-'' The .NET Foundation licenses this file to you under the MIT license.
-'' See the LICENSE file in the project root for more information.
-
-'Imports Microsoft.CodeAnalysis.CodeFixes
-'Imports Microsoft.CodeAnalysis.Diagnostics
-'Imports Microsoft.CodeAnalysis.VisualBasic.CodeFixes.Async
-
-'Namespace Microsoft.CodeAnalysis.Editor.VisualBasic.UnitTests.Diagnostics.Async
-'    Public Class AddAwaitTests
-'        Inherits AbstractVisualBasicDiagnosticProviderBasedUserDiagnosticTest
-
-'        Friend Overrides Function CreateDiagnosticProviderAndFixer(workspace As Workspace) As (DiagnosticAnalyzer, CodeFixProvider)
-'            Return (Nothing, New VisualBasicAddAwaitCodeFixProvider())
-'        End Function
-'    End Class
-'End Namespace
